[kernel XHCI]: refactor
The commit separates XHCI register handling into its own struct and
removes
unnecessary mutex usage from core XHCI components. Previously shared
state was
protected by a single mutex, which has been split into individual
mutexes per
component.
