+++ crates/kernel/src/drivers/xhci/mod.rs
@@ -2,9 +2,8 @@ use super::{
     interrupts::IRQInfo,
     utils::{read_ref, write_ref},
 };
-use regs::{CapsReg, OperationalRegs, RuntimeRegs, USBCmd, USBSts, XHCIDoorbellManager, XHCIIman};
+use regs::{CapsReg, XHCIDoorbellManager};
 use rings::{TRBCommand, XHCICommandRing, XHCIEventRing};
-use utils::allocate_buffers_frame;
 
 use crate::{
     arch::paging::current_higher_root_table,
@@ -12,16 +11,10 @@ use crate::{
     drivers::{
         interrupts::{self, IntTrigger, InterruptReceiver},
         pci::PCICommandReg,
-        xhci::rings::TRB_TYPE_ENABLE_SLOT_CMD,
+        xhci::{regs::XHCIRegisters, rings::TRB_TYPE_ENABLE_SLOT_CMD},
     },
-    memory::{
-        align_up,
-        frame_allocator::{self, Frame},
-        paging::{EntryFlags, PAGE_SIZE},
-    },
-    time,
+    memory::paging::{EntryFlags, PAGE_SIZE},
     utils::locks::Mutex,
-    PhysAddr,
 };
 
 use super::pci::PCIDevice;
@@ -32,214 +25,36 @@ mod utils;
 /// The maximum number of TRBs a CommandRing can hold
 const MAX_TRB_COUNT: usize = 256;
 
+// TODO: maybe stack interrupt stuff together in one struct behind a Mutex?
+/// The main XHCI driver Instance
 #[derive(Debug)]
 pub struct XHCI<'s> {
-    caps_regs: *mut CapsReg,
-    // TODO: free the frames when this goes out of scope? except that currently it never does
-    /// used to store the scratchpad_buffers pointers and the dcbaa (scratchpad_buffers, dcbaa)
-    buffers_frame: Frame,
-    scratchpad_buffers: Option<&'s mut [Frame]>,
-    dcbaa: &'s mut [PhysAddr],
-
-    command_ring: XHCICommandRing<'s>,
-    event_ring: XHCIEventRing<'s>,
-    doorbell_manager: XHCIDoorbellManager<'s>,
+    // TODO: maybe use a UnsafeCell here?
+    /// be careful using the registers, should only be used while interrupts are disabled
+    regs: Mutex<XHCIRegisters<'s>>,
+    /// Not accessed by interrupts
+    command_ring: Mutex<XHCICommandRing<'s>>,
+    /// Only accessed by interrupts
+    event_ring: Mutex<XHCIEventRing<'s>>,
+    /// Not accessed by interrupts
+    doorbell_manager: Mutex<XHCIDoorbellManager<'s>>,
     irq_info: IRQInfo,
 }
 
-impl<'s> InterruptReceiver for Mutex<XHCI<'s>> {
+impl<'s> InterruptReceiver for XHCI<'s> {
     fn handle_interrupt(&self) {
-        let mut this = self.lock();
-        let events = this.event_ring.dequeue_events();
-        crate::serial!("{events:#?}\n");
-
-        this.acknowledge_irq(0);
+        let events = self.event_ring.lock().dequeue_events();
+        crate::serial!("{events:#x?}\n");
+        self.regs.lock().acknowledge_irq(0);
     }
 }
 
-impl<'s> XHCI<'s> {
-    fn captabilities<'a>(&self) -> &'a CapsReg {
-        unsafe { &*self.caps_regs }
-    }
-
-    fn captabilities_mut<'a>(&mut self) -> &'a mut CapsReg {
-        unsafe { &mut *self.caps_regs }
-    }
-
-    fn operational_regs<'a>(&mut self) -> &'a mut OperationalRegs {
-        let caps = self.captabilities_mut();
-        caps.operational_regs_mut()
-    }
-
-    fn runtime_regs<'a>(&mut self) -> &'a mut RuntimeRegs {
-        self.captabilities_mut().runtime_regs_mut()
-    }
-
-    /// Clear any incoming interrupts for the interrupter
-    pub fn acknowledge_irq(&mut self, interrupter: u8) {
-        let op_regs = self.operational_regs();
-        // Write the USBSts::EINT bit to clear it, it is RW1C meaning write 1 to clear
-        write_ref!(op_regs.usbstatus, USBSts::EINT);
-
-        let runtime_regs = self.runtime_regs();
-        let interrupt_reg = runtime_regs.interrupter_mut(interrupter as usize);
-        // Similariy we clear the iman interrupt pending bit by writing 1 to it
-        let iman = interrupt_reg.iman | XHCIIman::INTERRUPT_PENDING;
-        write_ref!(interrupt_reg.iman, iman);
-    }
-
-    fn start(&mut self) {
-        let regs = self.operational_regs();
-        write_ref!(
-            regs.usbcmd,
-            regs.usbcmd | USBCmd::RUN | USBCmd::INTERRUPT_ENABLE
-        );
-
-        let timeout = 1000;
-        let time = time!();
-
-        while read_ref!(regs.usbstatus).contains(USBSts::HCHALTED) {
-            let now = time!();
-            if now >= time + timeout {
-                panic!(
-                    "timeout after {}ms while resetting the XHCI, HCHALTED did not clear: {:?}",
-                    now,
-                    read_ref!(regs.usbstatus)
-                )
-            }
-            core::hint::spin_loop();
-        }
-
-        assert!(!read_ref!(regs.usbstatus).contains(USBSts::NOT_READY));
-    }
-
-    #[allow(unused_unsafe)]
-    /// Resets the XHCI controller
-    fn reset(&mut self) {
-        let regs = self.operational_regs();
-
-        write_ref!(regs.usbcmd, regs.usbcmd & !USBCmd::RUN);
-
-        let timeout = 200;
-        let time = time!();
-
-        while !read_ref!(regs.usbstatus).contains(USBSts::HCHALTED) {
-            let now = time!();
-            if now >= time + timeout {
-                panic!(
-                    "timeout after {}ms while resetting the XHCI, HCHALTED did not set: {:?}",
-                    now,
-                    read_ref!(regs.usbstatus)
-                )
-            }
-            core::hint::spin_loop();
-        }
-
-        // reset the controller
-        write_ref!(regs.usbcmd, read_ref!(regs.usbcmd) | USBCmd::HCRESET);
-
-        let timeout = 1000;
-        let time = time!();
-
-        while read_ref!(regs.usbcmd).contains(USBCmd::HCRESET)
-            || read_ref!(regs.usbstatus).contains(USBSts::NOT_READY)
-        {
-            let now = time!();
-            if now >= time + timeout {
-                panic!(
-                    "timeout after {}ms while resetting controller, controller was never ready: {:?}",
-                    now - time,
-                    read_ref!(regs.usbcmd),
-                )
-            }
-            core::hint::spin_loop();
-        }
-        // asserts the controller was reset
-        assert_eq!(regs.usbcmd, USBCmd::empty());
-        assert_eq!(regs.dnctrl, 0);
-        assert_eq!(regs.crcr, 0);
-        assert_eq!(regs.dcbaap, PhysAddr::null());
-        assert_eq!(regs.config, 0);
-        write_ref!(regs.config, self.captabilities().max_device_slots() as u32);
-        // reconfigure the controller
-        self.reconfigure();
-        debug!(
-            XHCI,
-            "XHCI Reset\n{}\n{:#x?}",
-            regs,
-            self.runtime_regs().interrupter(0)
-        );
-    }
-
-    fn reconfigure(&mut self) {
-        let op_regs = self.operational_regs();
-        // Enable device notifications
-        write_ref!(op_regs.dnctrl, 0xFFFF);
-        self.configure_dcbaa();
-        self.configure_crcr();
-
-        self.configure_runtime();
-    }
-
-    fn configure_crcr(&mut self) {
-        let op_regs = self.operational_regs();
-        write_ref!(
-            op_regs.crcr,
-            *self.command_ring.base_phys_addr() | self.command_ring.current_ring_cycle() as usize
-        );
-    }
-
-    fn configure_dcbaa(&mut self) {
-        let caps = self.captabilities();
-        let op_regs = self.operational_regs();
-
-        // Allocates and sets the dcbaa
-        assert!(caps.max_device_slots() * size_of::<PhysAddr>() <= PAGE_SIZE);
-
-        let (dcbaa_slice, dcbaa_phys_addr) =
-            allocate_buffers_frame::<PhysAddr>(self.buffers_frame, 0, caps.max_device_slots());
-
-        // Allocates the scratchpad buffers array if neccassary
-        if caps.max_scratchpad_buffers() > 0 {
-            // uses the same frame to store the scratchpad_buffers pointers that we used to store dcbaa entries
-            // it is safe to do so as the max number of dcbaa entries is 255,
-            // and the max numbers of scratchpad_buffers is 15, (255 + 15) * 8 is very much less then the maximum amount of bytes a frame (page) can hold (4096)
-            // DCBAA entries must be 64 byte aligned
-            let (scratchpad_buffers, scratchpad_buffers_addr) = allocate_buffers_frame::<Frame>(
-                self.buffers_frame,
-                align_up((dcbaa_phys_addr + dcbaa_slice.len()).into_raw(), 64),
-                caps.max_scratchpad_buffers(),
-            );
-
-            for phys_addr in scratchpad_buffers.iter_mut() {
-                *phys_addr = frame_allocator::allocate_frame()
-                    .expect("XHCI: failed to allocate a page for a scratchpad buffer");
-            }
-            self.scratchpad_buffers = Some(scratchpad_buffers);
-            // DCBAA[0] is used to store the address of the scratchpad_buffers
-            self.dcbaa[0] = scratchpad_buffers_addr;
-        }
-
-        self.dcbaa = dcbaa_slice;
-        write_ref!(op_regs.dcbaap, dcbaa_phys_addr);
-    }
-
-    fn configure_runtime(&mut self) {
-        self.event_ring.reset();
-        let runtime_regs = self.runtime_regs();
-        let interrupt_reg = runtime_regs.interrupter_mut(0);
-        // Enable interrupts
-        write_ref!(interrupt_reg.iman, XHCIIman::INTERRUPT_ENABLE);
-
-        // Clear any pending interrupts
-        self.acknowledge_irq(0);
-    }
-}
+impl<'s> XHCI<'s> {}
 
 unsafe impl<'s> Send for XHCI<'s> {}
 unsafe impl<'s> Sync for XHCI<'s> {}
 
-impl<'s> PCIDevice for Mutex<XHCI<'s>> {
+impl<'s> PCIDevice for XHCI<'s> {
     fn class() -> (u8, u8, u8) {
         (0xc, 0x3, 0x30)
     }
@@ -276,7 +91,12 @@ impl<'s> PCIDevice for Mutex<XHCI<'s>> {
         let interrupter = runtime_regs.interrupter_mut(0);
 
         let command_ring = XHCICommandRing::create(MAX_TRB_COUNT);
-        let event_ring = XHCIEventRing::create(MAX_TRB_COUNT, interrupter);
+        let mut event_ring = XHCIEventRing::create(MAX_TRB_COUNT, interrupter);
+
+        let mut xhci_registers = unsafe { XHCIRegisters::new(caps_regs) };
+        unsafe {
+            xhci_registers.reconfigure(&mut event_ring, &command_ring);
+        }
 
         let doorbell_manager =
             XHCIDoorbellManager::new(caps_regs.doorbells_base(), caps_regs.max_device_slots());
@@ -287,36 +107,32 @@ impl<'s> PCIDevice for Mutex<XHCI<'s>> {
             .map(|msix| msix.into_irq_info())
             .unwrap();
 
-        let mut results = Mutex::new(XHCI {
-            buffers_frame: frame_allocator::allocate_frame()
-                .expect("XHCI: failed to allocate memory"),
-            scratchpad_buffers: None,
-            dcbaa: &mut [],
-            command_ring,
-            event_ring,
-            caps_regs,
-            doorbell_manager,
+        let mut this = XHCI {
+            command_ring: Mutex::new(command_ring),
+            event_ring: Mutex::new(event_ring),
+            doorbell_manager: Mutex::new(doorbell_manager),
+            regs: Mutex::new(xhci_registers),
             irq_info,
-        });
-        let this = results.get_mut();
+        };
         debug!(
             XHCI,
-            "Mapped\n{}\n{}",
-            this.captabilities(),
-            this.operational_regs()
+            "Created\n{}\n{}",
+            this.regs.get_mut().captabilities(),
+            this.regs.get_mut().operational_regs()
         );
-        this.reset();
-        results
+        this
     }
 
     fn start(&'static self) -> bool {
-        let irq_info = self.lock().irq_info.clone();
+        let irq_info = self.irq_info.clone();
         interrupts::register_irq(irq_info, IntTrigger::Edge, self);
 
-        let op_regs = self.lock().operational_regs();
+        let op_regs = self.regs.lock().operational_regs();
         let usbsts_before = read_ref!(op_regs.usbstatus);
         let usbcmd_before = read_ref!(op_regs.usbcmd);
-        self.lock().start();
+        unsafe {
+            self.regs.lock().start();
+        }
         let usbsts_after = read_ref!(op_regs.usbstatus);
         let usbcmd_after = read_ref!(op_regs.usbcmd);
         debug!(
@@ -329,8 +145,8 @@ impl<'s> PCIDevice for Mutex<XHCI<'s>> {
             0,
             0,
         );
-        self.lock().command_ring.enqueue(trb);
-        self.lock().doorbell_manager.ring_command_doorbell();
+        self.command_ring.lock().enqueue(trb);
+        self.doorbell_manager.lock().ring_command_doorbell();
         true
     }
 }
